Move Snippets to separate file
